+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -292,7 +292,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KRG5HGPYM5ZA3GZHXMSGDFYM"
+    id: Optional[str] = "01KRGC9RM6PR9S58NXPT0ZAW19"
 
     description: Optional[str] = None
 
@@ -1357,7 +1357,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KRG5HGPTBF87VPPDF48M3HNG"
+        "tool_01KRGC9RKS847BRKK3P3JWT1D1"
     )
 
     display_name: Optional[str] = None
@@ -1479,7 +1479,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KRG5HGPTNJ9BAS9451BQB03M"
+    id: Optional[str] = "01KRGC9RKRTK7TD27KSHV4DPXR"
 
     description: Optional[str] = None
 
@@ -1623,7 +1623,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KRG5HGPSPG02TF1BZJWZJ8D0"
+        "tool_01KRGC9RKQ3JND8K04NP5RZW67"
     )
 
     display_name: Optional[str] = None
@@ -1914,7 +1914,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KRG5HGPQQW3K4RK9XA93TPH2"
+        "tool_01KRGC9RKNT8SFNPV26D293Y2D"
     )
 
     display_name: Optional[str] = None
@@ -2099,7 +2099,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KRG5HGPP5VA3T2YWEYJSX01D"
+        "tool_01KRGC9RKM93EQ66YNF4C4WHJH"
     )
 
     display_name: Optional[str] = None
@@ -2288,7 +2288,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KRG5HGPN9YSQP1MDBD1ZGB2V"
+        "tool_01KRGC9RKKC1BG283XG4GQ71J8"
     )
 
     display_name: Optional[str] = None