+++ packages/orq-rc/RELEASES.md
@@ -7869,3 +7869,13 @@ Based on:
 - [python v4.10.0-rc.0] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.0] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.0 - packages/orq-rc
+
+## 2026-05-13 10:00:03
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.1] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.1] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.1 - packages/orq-rc